+++ envinorma/from_legifrance/legifrance_to_am.py
@@ -9,7 +9,9 @@ import bs4
 from bs4 import BeautifulSoup
 from leginorma import ArticleStatus, LegifranceArticle, LegifranceSection, LegifranceText
 
-from envinorma.data import (
+from envinorma.from_legifrance.numbering_exceptions import EXCEPTION_PREFIXES, MAX_PREFIX_LEN
+from envinorma.io.parse_html import extract_table
+from envinorma.models import (
     ArreteMinisteriel,
     EnrichedString,
     Link,
@@ -19,8 +21,6 @@ from envinorma.data import (
     TableReference,
     standardize_title_date,
 )
-from envinorma.from_legifrance.numbering_exceptions import EXCEPTION_PREFIXES, MAX_PREFIX_LEN
-from envinorma.io.parse_html import extract_table
 from envinorma.structure import split_alineas_in_sections
 from envinorma.title_detection import NumberingPattern, detect_patterns_if_exists, is_mainly_upper, is_probably_title
